+++ fast_rx/test/rx/rx_value_test.dart
@@ -1,7 +1,16 @@
 import 'package:fast_rx/fast_rx.dart';
+import 'package:fast_rx_test/fast_rx_test.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
+  test('RxValue registration', () {
+    expectRxRegistration<RxString>(
+      shouldRegister: [
+        RxTest(''.rx, (rx) => rx.value),
+      ],
+    );
+  });
+
   test('RxValue<int> notifications', () {
     final rx = 0.rx;