+++ fast_rx_test/lib/src/expect_rx_registration.dart
@@ -2,22 +2,28 @@
 
 import 'package:fast_rx/src/rx/rx.dart';
 import 'package:fast_rx/src/rx_notifier.dart';
+import 'package:fast_rx_test/fast_rx_test.dart';
 import 'package:fast_rx_test/src/fast_rx_testing.mocks.dart';
 import 'package:mockito/mockito.dart';
 
-/// Expect that every call in [shouldRegister] registers exactly once, and that
-/// every call in [shouldNotRegister] does not register
-expectRxRegistration(
-  Rx rx, {
-  required List<void Function()> shouldRegister,
-  List<void Function()> shouldNotRegister = const [],
+/// Expect that every test in [shouldRegister] registers exactly once, and that
+/// every test in [shouldNotRegister] does not register
+expectRxRegistration<T extends Rx>({
+  required List<RxTest<T>> shouldRegister,
+  List<RxTest<T>> shouldNotRegister = const [],
 }) {
   final notifier = MockRxNotifier();
   RxNotifier.instance = notifier;
 
-  for (final call in (shouldRegister + shouldNotRegister)) {
-    call();
+  for (final test in shouldRegister) {
+    test.transform(test.rx);
   }
 
-  verify(notifier.addStream(rx.stream)).called(shouldRegister.length);
+  verify(notifier.addStream(any)).called(shouldRegister.length);
+
+  for (final test in shouldNotRegister) {
+    test.transform(test.rx);
+  }
+
+  verifyNoMoreInteractions(notifier);
 }